feat(multio): Option to copy initial state diagnostics from grib (#36)
### Description
To be used by operations. AIFS does not output diagnostics at step 0,
but there's an operational requirement to do so.

An extra grib file with the missing diagnostics at step 0 can be
provided, and the fields in this file will be copied to the output. The
reason for doing it this way is so that all AIFS output goes through the
same multio pipeline.

### Contributor Declaration

By opening this pull request, I affirm the following:

* All authors agree to the [Contributor License
Agreement](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md).
* The code follows the project's coding standards.
* I have performed self-review and added comments where needed.
* I have added or updated tests to verify that my changes are effective
and functional.
* I have run all existing tests and confirmed they pass.